requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
google-generativeai>=0.3.0
openai>=1.0.0